% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAnalysis_New.R
\name{applyPCA}
\alias{applyPCA}
\title{Function that applies PCA on the preprocessed data and produces plots.}
\usage{
applyPCA(analysisList)
}
\arguments{
\item{analysisList}{list containing the Race column (from the original dataset in \code{dataset_file_path}) 
and the preprocessed, normalized, continuous dataset \code{normalizedContinuousDataset}}
}
\value{
a plot made with \cdoe{ggplot2} after applying PCA on \code{normalizedContinuousDataset}
}
\description{
Function that applies PCA on the preprocessed data and produces plots.
}
