% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSeparation_New.R
\name{check.integer}
\alias{check.integer}
\title{Function that checks if a provided variable is an integer
Helper function for \code{checkinteger.column}}
\usage{
check.integer(n)
}
\arguments{
\item{n}{data variable}
}
\value{
boolean value - true if the variable (\code{n}) is an integer, false if not
}
\description{
Function that checks if a provided variable is an integer
Helper function for \code{checkinteger.column}
}
\examples{
check.integer(1)
check.integer(12.34)
}
