% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSeparation_New.R
\name{checkinteger.column}
\alias{checkinteger.column}
\title{Function that checks if all the data in a column of a dataset are integers
Helper function for \code{getDiscreteAndContinuousIndices}}
\usage{
checkinteger.column(n)
}
\arguments{
\item{n}{a column of a dataframe}
}
\value{
a boolean value - true if the column has only integer values, false if not
}
\description{
Function that checks if all the data in a column of a dataset are integers
Helper function for \code{getDiscreteAndContinuousIndices}
}
