% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing_New.R
\name{dropColumnsWithMissingData}
\alias{dropColumnsWithMissingData}
\title{Function that drops columns in the discrete and continuous datasets which have missing data percentage
above a provided threshold}
\usage{
dropColumnsWithMissingData(missingDataThreshold, datasetsList_old)
}
\arguments{
\item{missingDataThreshold}{the threshold of missing data in a column for which that data should be removed}

\item{datasetsList_old}{list of the discrete and continuous datasets}
}
\value{
a similar list like the parameter passed in, with columns having too much missing data removed
}
\description{
Function that drops columns in the discrete and continuous datasets which have missing data percentage
above a provided threshold
}
