% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing_New.R
\name{dropNonNumericalData}
\alias{dropNonNumericalData}
\title{Function that drops non numerical data (not necessary for analysis) from the datasets}
\usage{
dropNonNumericalData(datasetsList_old)
}
\arguments{
\item{datasetsList_old}{a list of the discrete and continuous datasets of the form 
list("discrete" = discrete_dataset, "continuous" = continuous_dataset) (to be obtained from \code{addIdentifyingInfo}
 in \code{DataSeparation_New.R})}
}
\value{
a similar list like the parameter passed in, with non-numerical data removed
}
\description{
Function that drops non numerical data (not necessary for analysis) from the datasets
}
