% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSeparation_New.R
\name{getDiscreteAndContinuousIndices}
\alias{getDiscreteAndContinuousIndices}
\title{Function that gets the indices of the dataframe columns that have discrete data, and those that have continuous data}
\usage{
getDiscreteAndContinuousIndices(dataset)
}
\arguments{
\item{dataset}{a dataframe}
}
\value{
a list of the discrete and continuous indices, of the form 
list("discrete" = discrete_indices, "continuous" = continuous_indices)
}
\description{
Function that gets the indices of the dataframe columns that have discrete data, and those that have continuous data
}
