% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipelineDriver_New.R
\name{loadData}
\alias{loadData}
\title{Function that loads data to be analyzed from the specified files}
\usage{
loadData(dataset_file_path)
}
\arguments{
\item{dataset_file_path}{Relative file path to the file (containing the dataset) to be analyzed}
}
\value{
dataframe containing the dataset in \code{dataset_file_path}
}
\description{
Function that loads data to be analyzed from the specified files
}
