% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipelineDriver_New.R
\name{mainDriver}
\alias{mainDriver}
\title{Function that is the main driver for the entire program that executes the entire pipeline}
\usage{
mainDriver(dataset_file_path, missingDataThreshold)
}
\arguments{
\item{dataset_file_path}{Relative file path to the file (containing the dataset) to be analyzed}

\item{missingDataThreshold}{the threshold of missing data in a colum/row for which that data should be removed}
}
\value{
Plot after PCA Analysis of the Preprocessed, Normalized (Continuous Data) Dataset
(obtained from \code{dataset_file_path})
}
\description{
Function that is the main driver for the entire program that executes the entire pipeline
}
