% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing_New.R
\name{makeDataTypesUniform}
\alias{makeDataTypesUniform}
\title{Function that converts all the data in both datasets to be of uniform data type (double)}
\usage{
makeDataTypesUniform(datasetsList_old)
}
\arguments{
\item{datasetsList_old}{a list of the discrete and continuous datasets of the form 
list("discrete" = discrete_dataset, "continuous" = continuous_dataset)}
}
\value{
a similar list like the parameter passed in, with uniform data types
}
\description{
Function that converts all the data in both datasets to be of uniform data type (double)
}
