% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing_New.R
\name{missingPercent}
\alias{missingPercent}
\title{Function that gets the percentage of missing data in the columns of a dataset
Helper function for \code{dropColumnsWithMissingData}}
\usage{
missingPercent(dataset)
}
\arguments{
\item{dataset}{a dataframe}
}
\value{
a list of the percentages of missing data in every column
}
\description{
Function that gets the percentage of missing data in the columns of a dataset
Helper function for \code{dropColumnsWithMissingData}
}
