% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing_New.R
\name{normalizeCol}
\alias{normalizeCol}
\title{Function that normalizes data in the column of a dataset
Helper function for \code{normalizeContinuousData}}
\usage{
normalizeCol(n)
}
\arguments{
\item{n}{a dataframe column}
}
\value{
the column, with data normalized
}
\description{
Function that normalizes data in the column of a dataset
Helper function for \code{normalizeContinuousData}
}
