% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing_New.R
\name{normalizeContinuousData}
\alias{normalizeContinuousData}
\title{Function that normalizes all the data in the continuous dataset}
\usage{
normalizeContinuousData(continuousDataset)
}
\arguments{
\item{continuousDataset}{the dataset containing continuous data}
}
\value{
the dataset, normalized
}
\description{
Function that normalizes all the data in the continuous dataset
}
